% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generes_to_heatmap.R
\name{}
\alias{}
\alias{generes_to_heatmap}
\title{Generate signature matrix}
\usage{
generes_to_heatmap(generes, species = "human", naming_preference = -9,
  rda_path = "~/scMappR/data", make_names = T, internal = FALSE)
}
\arguments{
\item{generes}{A list of cell-tpe markers with fold-changes and P-vlaues (FindMarkers output in Seurat)}

\item{species}{The species of gene symbols, if not internal, "human" or "mouse"}

\item{naming_preference}{Likely cell-types given tissues (to be passed into human_mouse_ct_marker_enrich)}

\item{make_names}{Identify names of cell-type markers using the Fisher's Exact-Test method (T/F).}

\item{internal}{If this function is pre-processing from panglao (T/F).}
}
\value{
\code{generes_to_heatmap} A list containing a signature matrix by rank := -1*log10(Pfdr) and by fold-change (only increasing). Additionally it returns the top (up to) 30 CT markers for each cell-type, as well as the name of each cell-type (from the signature methods method). \cr
}
\description{
Convert a list of cell-type markers from FindMarkers in Seurat to a signature matrix defined by odds ratio and rank.
}
\details{
Take an list of compiled DEGs from different cell types, identify what the cell-types are using the fisher's exact method, and then convert into a signature matrix for both the adjusted p-value and odds ratio
}
\examples{
 
 
 # load generes object
load("~/scMappR/data/Preoptic_region_example.rda")
#data(Preoptic_region_example)
signature <- generes_to_heatmap(POA_generes,species = -9, make_names = F)

}
