% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_generation--fileChange.R
\name{heatmap_generation}
\alias{heatmap_generation}
\title{Generate Heatmap}
\usage{
heatmap_generation(genesIn, comp, cex = 0.8, cellTypes = "ALL",
  pVal = 0.01, isPval = TRUE, isMax = F, isBackground = F,
  reference = "C:/Users/Dustin Sokolowski/Desktop/romanov_wilcoxon_test_2.RData",
  which_species = "human")
}
\arguments{
\item{genesIn}{A list of gene symbols (all caps) to have their cell type enrichment}

\item{comp}{The name of the comparison}

\item{cellTypes}{Colnames of the cell-types you will extract (passed to extract_genes_cell)}

\item{pVal}{The level of association a gene has within a cell type (passed to extract_genes_cell)}

\item{isMax}{If you are taking the single best CT marker (T/F) -- TRUE not reccomended}

\item{isBackground}{If the heatmap is from the entire signature matrix or just the inputted gene list (T/F). isBackground == TRUE is used for internal.}

\item{which_species}{Species of gene symbols -- "human" or "mouse"}

\item{Cex}{the size of the genes in the column label for the heatmap}

\item{isPvalue}{If the signature matrix is raw p-value (T/F) -- TRUE not reccomended}

\item{isMax}{only take the top most DEG for each cell-type (T/F) -- TRUE not reccomended}

\item{refence}{Path to signature matrix or the signature matrix itself.}
}
\value{
\code{heatmap_generation} A heatmap/barplot of p-value or odds-ratio of cell-type specific genes intersecting with the gene list. A list of genes that do/don't intersect with the signature matrix as well as a list of which cell-type these over-represented genes live in. \cr
}
\description{
This function takes an inputted signature matrix as well as a list of genes and overlaps them.
 Then, if there is overlap, it prints a heatmap or barplot (depending on the number of overlapping genes)
 Then, for every cell-type, genes considered over-represented are saved in a list
}
\examples{

# load in signature matrices
load("data/Preoptic_region_example.rda")
# data(Preoptic_region_example)
Signature <- POA_Rank_signature
 rowname <- get_gene_symbol(Signature)
 rownames(Signature) <- rowname$rowname
 genes <- rownames(Signature)[1:100]
 heatmap_test <- heatmap_generation(genesIn = genes, "scMappR_test", reference = Signature, which_species = "mouse")
 
}
