% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_TF_barplot.R
\name{make_TF_barplot}
\alias{make_TF_barplot}
\title{Plot g:ProfileR Barplot (TF)}
\usage{
make_TF_barplot(ordered_back_all_tf, top_tf = 5)
}
\arguments{
\item{top_tf}{The number of pathways you want to plot}

\item{ordered_back_all}{output of the g:ProfileR function}
}
\value{
\code{make_TF_barplot} A barplot of the number of "top_tf" tf names (not motifs), ranked by -log10(Pfdr). \cr
}
\description{
Make a barplot of the top transcription factors enriched by g:ProfileR
}
\details{
This function takes a g:ProfleR output and prints the top "top_tfs" most significantly
enriched P-values before plotting the rank of their P-values
}
\examples{
\\notrun {
library(gProfileR)
# load in signature matrices
load("~/scMappR/data/Preoptic_region_example.rda")
# data(Preoptic_region_example)
Signature <- as.data.frame(POA_Rank_signature)
rowname <- get_gene_symbol(Signature)
rownames(Signature) <- rowname$rowname
genes <- rownames(Signature)[Signature$` Enteroendocrine cell_and_Neuroendocrine cells_17` > 1]
BP <- gProfileR::gprofiler(genes, "mmusculus", src_filter = c("GO:BP", "KEGG", "REAC"), max_set_size = 2000, exclude_iea = T)
TF <- gProfileR::gprofiler(genes, "mmusculus", src_filter = c("TF"), max_set_size = 5000, exclude_iea = T)
bp <- plotBP(BP)
tf <- make_TF_barplot(TF)
 }
}
