% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tochr.R
\name{tochr}
\alias{tochr}
\title{To Character.}
\usage{
tochr(x)
}
\arguments{
\item{x}{a vector of character, factor, or numeric}
}
\value{
\code{tochr} Returns a character vector. \cr
}
\description{
This function checks if your vector is not a character and will then converts it to a character.
}
\examples{

 
 # vector of factors
 fact <- factor(c("a", "b", "c", "d"))
 # convert to character
 char <- tochr(fact)
 

}
